@@ -10,8 +10,9 @@ module GardenCrops
     optional do
       string :quantity
       string :guide_id
+      string :crop_id
       string :stage
-      string :sowed
+      date :sowed
     end
 
     def garden_crop
@@ -20,6 +21,7 @@ module GardenCrops
 
     def validate
       validate_guide
+      validate_crop
       validate_garden
       validate_permissions
     end
@@ -36,6 +38,15 @@ module GardenCrops
     rescue Mongoid::Errors::DocumentNotFound
       msg = "Could not find a guide with id #{guide_id}."
       add_error :guide, :guide_not_found, msg
+    end
+
+    def validate_crop
+      if crop_id
+        @crop = Crop.find(crop_id)
+      end
+    rescue Mongoid::Errors::DocumentNotFound
+      msg = "Could not find a crop with id #{crop_id}."
+      add_error :crop, :crop_not_found, msg
     end
 
     def validate_garden
@@ -55,9 +66,10 @@ module GardenCrops
     def set_params
       garden_crop.garden      = @garden
       garden_crop.guide       = @guide if @guide.present?
+      garden_crop.crop        = @crop if @crop.present?
       garden_crop.quantity    = quantity if quantity.present?
       garden_crop.stage       = stage if stage.present?
-      garden_crop.sowed       = sowed if stage.present?
+      garden_crop.sowed       = sowed if sowed.present?
       garden_crop.save
     end
   end
