@@ -1,4 +1,5 @@
 require 'spec_helper'
+require 'openfarm_errors'
 
 describe Gardens::DestroyGarden do
   let(:mutation) { Gardens::DestroyGarden }
@@ -19,12 +20,10 @@ describe Gardens::DestroyGarden do
   it 'checks that a garden with the required id exists' do
     params[:id] = 1
     outcome = mutation.run(params)
-    expect(outcome.success?).to be_false
+    expect(outcome.success?).to be_falsey
     expect(outcome.errors.message_list).to include('Could not find a garden' +
                                                    " with id #{params[:id]}.")
   end
 
-  # We can't test for mutations that include the openfarm error
-  # because it's defined outside of the scope of the standalone
-  # mutation?
+  # Test for things that use openfarm_errors
 end
