@@ -3,8 +3,8 @@
 FactoryGirl.define do
   factory :garden_crop do
     quantity    { rand(100) }
-    stage       { "#{Faker::Lorem.word}" }
-    sowed       { "#{Faker::Date.between(2.days.ago, Date.today)}" }
+    sowed       { Faker::Date.between(2.days.ago, Date.today) }
+    stage
     garden
     guide
   end
